Refactored to move elementRefs to within state, moving the module within the state dir.
