+++ client/listeners/resetListeners.js
@@ -1,4 +1,4 @@
-const { addLoadRef } = require('./../utils/elementRefs')
+const { addLoadRef } = require('./../state/elementRefs')
 const state = require('./../state/state.js')
 const addTrackForm = require('./../utils/addTrackForm.js')
